+++ R/2_4_0_textPredict_Assess_Classify.R
@@ -27,6 +27,8 @@
 #' @param dim_names (boolean; only for "text-trained"-models) Account for specific dimension names from textEmbed()
 #' (rather than generic names including Dim1, Dim2 etc.). If FALSE the models need to have been trained on
 #' word embeddings created with dim_names FALSE, so that embeddings were only called Dim1, Dim2 etc.
+#' @param check_matching_word_embeddings (boolean) If `TRUE`, the function will check whether the word embeddings (model type and layer) match
+#' the requirement of the trained model - if a mis-match is found the function till stop. If `FALSE`, the function will not verify.
 #' @param language_distribution (character column; only for "text-trained" models) If you provide the raw language data used for making the embeddings used for assessment,
 #' the language distribution (i.e., a word and frequency table) will be compared with saved one in the model object (if one exists).
 #' This enables calculating similarity scores.
@@ -148,6 +150,7 @@ textPredict <- function(
     x_append = NULL,
     append_first = TRUE,
     dim_names = TRUE,
+    check_matching_word_embeddings = TRUE,
     language_distribution = NULL,
     language_distribution_min_words = "trained_distribution_min_words",
     save_model = TRUE,
@@ -259,6 +262,7 @@ textPredict <- function(
         x_append = x_append,
         append_first = append_first,
         dim_names = dim_names,
+        check_matching_word_embeddings = check_matching_word_embeddings,
         language_distribution = language_distribution,
         language_distribution_min_words = language_distribution_min_words,
         save_model = save_model,